pyproject.toml
[project]
authors = [
  {name = "seonglae", email = "sungle3737@gmail.com"},
]
dependencies = [
    "fire>=0.5.0",
    "transformers>=4.34.0",
    "torch==2.1.0+cu118",
    "chromadb>=0.4.14",
    "datasets>=2.14.5",
    "teicli>=0.3.1",
    "faiss-cpu>=1.7.4",
    "vecs>=0.4.1",
]
description = "Add your description here"
name = "curiosity"
readme = "README.md"
requires-python = ">= 3.8"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.rye]
dev-dependencies = [
  "autopep8>=2.0.4",
  "pip>=23.2.1",
  "mypy>=1.6.0",
]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[[tool.rye.sources]]
name = "cuda"
type = "index"
url = "https://download.pytorch.org/whl/cu118"
